Final clean version

+++ ft_strdup.c
@@ -6,7 +6,7 @@
 /*   By: aperin <aperin@student.s19.be>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/04 16:00:11 by aperin            #+#    #+#             */
-/*   Updated: 2022/10/05 15:47:32 by aperin           ###   ########.fr       */
+/*   Updated: 2022/10/22 17:25:33 by aperin           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,10 @@
 
 char	*ft_strdup(const char *str)
 {
-	int		len;
+	size_t	len;
 	char	*dup;
 
-	len = (int) ft_strlen(str);
+	len = ft_strlen(str);
 	dup = malloc((len + 1) * sizeof(char));
 	if (!dup)
 		return (0);
